update(update-notifier): v6 (ESM rewrite)
Update notifier is now ESM module. It's dependencies are also ESM
modules. The direct dependency tree, including type-fest utils library,
required bump to minumum TS version 4.1 and bump in `lib` definition, as
DT cannot use skip library check option.

- ESM module rewrite
- reflect actual module naming and structure

https://github.com/yeoman/update-notifier/releases/tag/v6.0.0

Thanks!
